requirements.txt
django
djangorestframework
markdown
django-filter
pygments
djoser
drf_yasg
pytest
pytest-django
black
flake8